It’s possible to generate more than TOTAL_SUPPLY of tokens
We are not having a TOTAL_SUPPLY anymore. We just have a
FOUNDATION_SUPPLY that is the amount of tokens that will minted and
tranferred to the qiibee wallet in the finalize() function.

The total supply will be kind of "controlled" by the rate with which
the crowdsale has been initialized.